Autentificacion LDAP

+++ app/api/tasks/route.ts
@@ -1,7 +1,15 @@
 import { NextResponse } from 'next/server';
 import pool from '@/lib/db';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/auth';
 
 export async function GET(request: Request) {
+  // const session = await getServerSession(authOptions);
+  // console.log("[Tasks API] Session:", session)
+  // if (!session) {
+  //   return NextResponse.json({ error: 'No autorizado' }, { status: 401 });
+  // }
+  
   const { searchParams } = new URL(request.url);
   const mode = searchParams.get('mode');
 
@@ -63,6 +71,11 @@ export async function GET(request: Request) {
 }
 
 export async function POST(request: Request) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'No autorizado' }, { status: 401 });
+  }
+  
   try {
     const data = await request.json();
     
@@ -102,6 +115,11 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'No autorizado' }, { status: 401 });
+  }
+  
   try {
     const { id } = await request.json();
     await pool.query('DELETE FROM tasques WHERE id = ?', [id]);
